requirements.txt
# Django and REST Framework
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0

# LangChain and LangGraph
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0

# Database
dj-database-url>=2.1

# Redis (for conversation memory)
redis>=5.0
django-redis>=5.4

# OpenAI
openai>=1.30

# Utilities
pydantic>=2.0
python-dateutil>=2.8

# API Documentation
drf-spectacular>=0.27

# Development
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
httpx>=0.27

# Data Generation
faker>=24.0

# Async Support
uvicorn>=0.29
gunicorn>=21.0

# Type hints
typing-extensions>=4.10
